Task2
add walls and adjust the moving area of player accordingly. Since my computer screen is not large enough to view the whole 1000*1000 screen so I changed the size of screen and walls a bit. I declared the size of walls and screen in two tuples at the begining of the code. you can chage the size of them both if you want to. The default is 500*500 screen and 10*10 walls